# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_html_version_api_v1_presentations__presentation_id__versions_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_html_version_api_v1_presentations__presentation_id__versions_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SlideRead": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "selector": {
            "title": "Selector",
            "type": "string"
          },
          "slide_index": {
            "title": "Slide Index",
            "type": "integer"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "id",
          "slide_index",
          "selector",
          "title"
        ],
        "title": "SlideRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VersionRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "entry_path": {
            "title": "Entry Path",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "presentation_id": {
            "format": "uuid",
            "title": "Presentation Id",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "slides": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SlideRead"
            },
            "title": "Slides",
            "type": "array"
          },
          "storage_kind": {
            "title": "Storage Kind",
            "type": "string"
          },
          "version_number": {
            "title": "Version Number",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "presentation_id",
          "version_number",
          "origin",
          "storage_kind",
          "entry_path",
          "sha256",
          "size_bytes",
          "created_at"
        ],
        "title": "VersionRead",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Internal presentation collaboration API (v1 scaffold).",
    "title": "promptDeck API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/presentations/{presentation_id}/versions": {
      "get": {
        "operationId": "list_versions_api_v1_presentations__presentation_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "presentation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Presentation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VersionRead"
                  },
                  "title": "Response List Versions Api V1 Presentations  Presentation Id  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Versions",
        "tags": [
          "versions"
        ]
      },
      "post": {
        "operationId": "upload_html_version_api_v1_presentations__presentation_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "presentation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Presentation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_html_version_api_v1_presentations__presentation_id__versions_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Html Version",
        "tags": [
          "versions"
        ]
      }
    },
    "/api/v1/presentations/{presentation_id}/versions/{version_id}/activate": {
      "post": {
        "operationId": "activate_version_api_v1_presentations__presentation_id__versions__version_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "presentation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Presentation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Activate Version",
        "tags": [
          "versions"
        ]
      }
    },
...
  }
}